split out options handler
